@@ -5,7 +5,7 @@ publish_to: none
 
 environment:
   sdk: ">=3.3.1 <4.0.0"
-  flutter: ^3.19.3
+  flutter: ^3.22.0
 
 dependencies:
   cached_network_image: ^3.3.1
@@ -15,7 +15,7 @@ dependencies:
   flutter_cache_manager: ^3.3.1
   flutter_svg: ^2.0.9
   go_router: ^13.2.4
-  intl: ">=0.18.0 <0.19.0"
+  intl: ">=0.18.0 <=0.19.0"
   pausable_timer: ^3.1.0+3
   shared:
     path: ../shared
