@@ -6,14 +6,14 @@ publish_to: none
 
 environment:
   sdk: ">=3.3.1 <4.0.0"
-  flutter: ^3.19.3
+  flutter: ^3.22.0
 
 dependencies:
   equatable: ^2.0.5
   flutter:
     sdk: flutter
   json_annotation: ^4.8.1
-  meta: ">=1.10.0 <1.12.0"
+  meta: ">=1.10.0 <=1.12.0"
   shared:
     path: ../shared
   user_repository:
