@@ -5,7 +5,7 @@ publish_to: none
 
 environment:
   sdk: ">=3.3.1 <4.0.0"
-  flutter: ^3.19.3
+  flutter: ^3.22.0
 
 dependencies:
   firebase_messaging: ^14.7.10
